Move fetch constants under the shared folder
